+++ Controllers/SaleController.cs
@@ -1,8 +1,10 @@
-﻿using CarQuery__Test.Domain.Models;
+﻿using CarQuery__Test.Authentication.Services;
+using CarQuery__Test.Domain.Models;
 using CarQuery__Test.Domain.Services;
 using CarQuery__Test.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace CarQuery__Test.Controllers
 {
@@ -58,6 +60,13 @@ namespace CarQuery__Test.Controllers
         [HttpPost] //Create a new Sale
         public async Task<IActionResult> CreateSale([FromBody] Sale sale)
         {
+            string role = TokenService.GetValueFromClaim(HttpContext.User.Identity, ClaimTypes.Role);
+
+            if (role.Equals("U"))
+            {
+                return BadRequest($"User without access");
+            }
+
             try
             {
                 var createdSale = await _saleService.CreateSaleAsync(sale);
@@ -80,6 +89,12 @@ namespace CarQuery__Test.Controllers
         [HttpPut("{id}")] //Update a Sale
         public async Task<IActionResult> UpdateSale(int id, [FromBody] Sale sale)
         {
+            string role = TokenService.GetValueFromClaim(HttpContext.User.Identity, ClaimTypes.Role);
+
+            if (role.Equals("U"))
+            {
+                return BadRequest($"User without access");
+            }
 
             try
             {
@@ -99,6 +114,13 @@ namespace CarQuery__Test.Controllers
         [HttpDelete("{id}")] //Delete a Sale
         public async Task<IActionResult> DeleteSale(int id)
         {
+            string role = TokenService.GetValueFromClaim(HttpContext.User.Identity, ClaimTypes.Role);
+
+            if (role.Equals("S") || role.Equals("U"))
+            {
+                return BadRequest($"User without access");
+            }
+
             try
             {
                 var result = await _saleService.DeleteSaleAsync(id);
