+++ Controllers/UserController.cs
@@ -1,8 +1,10 @@
-﻿using CarQuery__Test.Domain.Models;
+﻿using CarQuery__Test.Authentication.Services;
+using CarQuery__Test.Domain.Models;
 using CarQuery__Test.Domain.Services;
 using CarQuery__Test.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace CarQuery__Test.Controllers
 {
@@ -83,6 +85,13 @@ namespace CarQuery__Test.Controllers
         [HttpPost] //Create a new user
         public async Task<IActionResult> CreateUser([FromBody] User user)
         {
+            string role = TokenService.GetValueFromClaim(HttpContext.User.Identity, ClaimTypes.Role);
+
+            if (role.Equals("S") || role.Equals("U"))
+            {
+                return BadRequest($"User without access");
+            }
+
             try
             {
                 var createdUser = await _userService.CreateUserAsync(user);
@@ -105,6 +114,12 @@ namespace CarQuery__Test.Controllers
         [HttpPut("{id}")] //Update a user
         public async Task<IActionResult> UpdateUser(int id, [FromBody] User user)
         {
+            string role = TokenService.GetValueFromClaim(HttpContext.User.Identity, ClaimTypes.Role);
+
+            if (role.Equals("S") || role.Equals("U"))
+            {
+                return BadRequest($"User without access");
+            }
 
             try
             {
@@ -124,6 +139,13 @@ namespace CarQuery__Test.Controllers
         [HttpDelete("{id}")] //Delete a user
         public async Task<IActionResult> DeleteUser(int id)
         {
+            string role = TokenService.GetValueFromClaim(HttpContext.User.Identity, ClaimTypes.Role);
+
+            if (role.Equals("S") || role.Equals("U"))
+            {
+                return BadRequest($"User without access");
+            }
+
             try
             {
                 var result = await _userService.DeleteUserAsync(id);
